feat: add PUT and DELETE endpoints with validations

+++ app/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Date
+from sqlalchemy import Column, Integer, String, Numeric, Date
 from app.database import Base
 import datetime
 
@@ -7,6 +7,6 @@ class Transaction(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     description = Column(String, nullable=False)
-    amount = Column(Float, nullable=False)
+    amount = Column(Numeric(10, 2), nullable=False)
     date = Column(Date, default=datetime.date.today)
     category = Column(String, nullable=False)
